tradutor_de_artigos_tecnicos_com_azure_ai/requirements.txt
python-docx==1.1.2
requests==2.32.3
beautifulsoup4==4.12.3
openai==1.53.0
langchain-openai==0.2.4
